requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# SQLite (for development)
# For PostgreSQL, also add: psycopg2-binary==2.9.9

# Authentication
pyjwt[crypto]==2.8.0
cryptography==41.0.7
requests==2.31.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# OpenAI
openai==1.3.7

# Stripe payments
stripe==7.8.0

# HubSpot (optional)
hubspot-api-client==7.5.0

# Email
emails==0.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging
structlog==23.2.0

# CORS and middleware
python-jose[cryptography]==3.3.0
